refactor: hoist inline imports and move ConditionType to rule_validation
Move function-level imports to module top-level in tagging and test
files. Relocate ConditionType enum to rule_validation.py so validators
can use the enum directly instead of string-literal constants.
