+++ apps/evaluations/models.py
@@ -15,6 +15,7 @@ from pydantic import BaseModel as PydanticBaseModel
 
 from apps.chat.models import ChatMessage, ChatMessageType
 from apps.evaluations.rule_validation import (
+    ConditionType,
     validate_condition,
     validate_field_in_schema,
     validate_tag_compatibility,
@@ -463,37 +464,6 @@ class EvaluationRunAggregate(BaseModel):
 
     class Meta:
         unique_together = ("run", "evaluator")
-
-
-class ConditionType(models.TextChoices):
-    EQUALS = "equals", "Equals"
-    RANGE = "range", "Range"
-
-    @staticmethod
-    def coerce_value(raw, field_type: str | None):
-        """Coerce a raw equals-value input to the target field's python type.
-
-        Returns the input unchanged when `field_type` isn't numeric; raises
-        (TypeError, ValueError) on failed int/float conversion so the form can
-        surface a user-facing error.
-        """
-        if field_type == "int":
-            return int(raw)
-        if field_type == "float":
-            return float(raw)
-        return raw
-
-    def matches(self, condition_value: dict, field_value) -> bool:
-        """Return True if field_value satisfies this condition."""
-        if self == ConditionType.EQUALS:
-            return field_value == condition_value.get("value")
-        if self == ConditionType.RANGE:
-            try:
-                numeric = float(field_value)
-            except (TypeError, ValueError):
-                return False
-            return float(condition_value["min"]) <= numeric <= float(condition_value["max"])
-        raise ValueError(f"Unknown condition type: {self}")
 
 
 class EvaluatorTagRule(BaseTeamModel):
